@@ -6,31 +6,30 @@ authors = ["Adam Žúrek <adamzurek14@gmail.com>", "Erik Belák <erik.belak@gmai
 license = "MIT"
 
 [tool.poetry.dependencies]
-python = "^3.9"
+python = "^3.12"
 python-dotenv = "^1.0.0"
 porcupine-python = "^0.4.0"
-django = "^3.2.1"
-django-filter = "^2.2.0"
-django-api-forms = "^1.0.0-rc.9"
+django = "^5.0"
+django-filter = "^24.2"
+django_api_forms = "1.0.0rc9"
 django-enum-choices = "^2.1.2"
 django-object-checker = "^1.0.0"
 django-camel-spitter = "^0.3.2"
 django-imap-backend = "^0.2.3"
 django-role-permissions = "^3.1.1"
-django-bootstrap-v5 = "^1.0.11"
 django-cors-headers = "^3.12.0"
 
 Markdown = "^3.4.1"
 psycopg2-binary = "^2.8.5"
 pycryptodome = "^3.9.8"
 pyopenssl = "^19.1.0"
-argon2-cffi = "^20.1.0"
+argon2-cffi = "^23.1.0"
 pyotp = "^2.4.0"
 qrcode = "^6.1.0"
 pillow = "^10.0"
 python-ldap = "^3.4.4"
 
-sentry-sdk = "^0.18.0"
+sentry-sdk = "^1.39.1"
 flake8 = "^7.0.0"
 
 [tool.poetry.dev-dependencies]
