@@ -8,6 +8,6 @@ from apps.core.models.token import Token
 from apps.core.models.language import Language
 from apps.core.models.api_key import ApiKey
 from apps.core.models.log_entry import LogEntry
-from apps.core.models.log import Log
+from apps.core.models.audit_log import AuditLog
 from apps.core.models.auth_source import AuthSource
 from apps.core.models.auth_source_group import AuthSourceGroup
